Restructure test class packages for consistency
