+++ dev/com.ibm.ws.jsfContainer_fat_2.2/fat/src/com/ibm/ws/jsf/container/fat/utils/JSFUtils.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2018 IBM Corporation and others.
+ * Copyright (c) 2018, 2020 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -8,7 +8,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.jsf.container.fat;
+package com.ibm.ws.jsf.container.fat.utils;
 
 import java.net.URL;
